requirements.txt
Flask==3.1.2
Werkzeug==3.1.4
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
google-generativeai==0.8.3
flake8==7.1.0
black==24.4.2
pytest==8.3.2
